1. Move dependency on setuptools closer to where it's needed 2. Scope multiprocessing Pool so a pool left open doesn't hang the process 3. Move run_setup_py to its own file so the subprocess doesn't need to import much
